server: replace (":") in blob names with ("-")
This commit fixes issues where manually moving blobs from one OS that
allows colons in file paths to another OS that does not allow colons.

The fix is to replace colons with dashes in the blob names on `ollama
serve` but before binding and serving the blobs, and then consistently
use the dash-replaced blob names throughout other server operations.

Fixes #2032